backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
gunicorn
